pyproject.toml
[project]
name = "glpi-universal-backend"
version = "0.1.0"
description = "Backend Universal para API GLPI - Camada centralizada para todas as aplicações"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "pybreaker>=1.3.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
